pyproject.toml
[project]
name = "ai-job-hunter"
version = "3.0.0"
description = "Daily ML/AI/Data Science job scraper with SQLite storage and Telegram notifications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.10",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.0",
    "redis>=5.2.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "ruff>=0.15.4",
    "seaborn>=0.13.2",
    "uvicorn>=0.36.0",
]

[tool.pytest.ini_options]
markers = [
    "live_network: opt-in tests that probe live public ATS endpoints",
]
